+++ NEWS.md
@@ -1,1 +1,3 @@
 * version **0.1.0** : original version with the basic functions used to estimate the radial velocity in stellar spectra
+
+* version **0.1.1** : fixed a couple small bugs in the *gaussfit( )* function to allow for easier parallelization